feat: move key-related traits to separate module
Move PublicKeyParts to the separate module.
