+++ src/pkcs1v15.rs
@@ -30,7 +30,8 @@ use crate::algorithms::pkcs1v15::*;
 use crate::dummy_rng::DummyRng;
 use crate::errors::{Error, Result};
 use crate::internals::{uint_to_be_pad, uint_to_zeroizing_be_pad};
-use crate::key::{self, PublicKeyParts};
+use crate::key;
+use crate::keytraits::PublicKeyParts;
 use crate::padding::{PaddingScheme, SignatureScheme};
 use crate::traits::{Decryptor, EncryptingKeypair, RandomizedDecryptor, RandomizedEncryptor};
 use crate::{RsaPrivateKey, RsaPublicKey};
